requirements.txt
google-genai
pyaudio
opencv-python
pillow
python-dotenv